tidying up src/
