+++ R/RcppExports.R
@@ -1,6 +1,18 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+rcpp_get_composition_vector <- function(x) {
+    .Call('_landscapemetrics_rcpp_get_composition_vector', PACKAGE = 'landscapemetrics', x)
+}
+
+rcpp_get_entropy <- function(x, base = "log2") {
+    .Call('_landscapemetrics_rcpp_get_entropy', PACKAGE = 'landscapemetrics', x, base)
+}
+
+lsm_unique <- function(x) {
+    .Call('_landscapemetrics_lsm_unique', PACKAGE = 'landscapemetrics', x)
+}
+
 #' Coordinates from a matrix
 #'
 #' This function gets coordinates (row and column numbers) of the matrix cells.
@@ -40,50 +52,12 @@ rcpp_create_neighborhood <- function(directions) {
     .Call('_landscapemetrics_rcpp_create_neighborhood', PACKAGE = 'landscapemetrics', directions)
 }
 
-#' Adjacent cells
-#'
-#' Identify cells that are adjacent to a set of cells on a matrix.
-#' An output is a two columns matrix, where the first column contains a
-#' main cell number and the second column contains an adjacent cell number.
-#'
-#' @param x A matrix
-#' @param directions The number of directions in which cells should be connected:
-#' 4 (rook's case), 8 (queen's case), or a neighbourhood matrix.
-#' The neighbourhood matrix should have one cell with value 0 (the focal cell),
-#' and at least one cell with value 1 (the adjacent cells).
-#' Cells with other values (e.g. NA) are ignored.
-#' @keywords internal
-rcpp_get_adjacency <- function(x, directions) {
-    .Call('_landscapemetrics_rcpp_get_adjacency', PACKAGE = 'landscapemetrics', x, directions)
-}
-
-#' Adjacent cells values
-#'
-#' Extract values of adjacent cells.
-#' An output is a two columns matrix, where the first column contains a
-#' main cell value and the second column contains an adjacent cell value.
-#'
-#' @param x A matrix
-#' @param directions The number of directions in which cells should be connected:
-#' 4 (rook's case), 8 (queen's case), or a neighbourhood matrix.
-#' The neighbourhood matrix should have one cell with value 0 (the focal cell),
-#' and at least one cell with value 1 (the adjacent cells).
-#' Cells with other values (e.g. NA) are ignored.
-#' @keywords internal
-rcpp_get_pairs <- function(x, directions) {
-    .Call('_landscapemetrics_rcpp_get_pairs', PACKAGE = 'landscapemetrics', x, directions)
-}
-
-rcpp_get_composition_vector <- function(x) {
-    .Call('_landscapemetrics_rcpp_get_composition_vector', PACKAGE = 'landscapemetrics', x)
+rcpp_get_circle <- function(points, resolution) {
+    .Call('_landscapemetrics_rcpp_get_circle', PACKAGE = 'landscapemetrics', points, resolution)
 }
 
 rcpp_get_coocurrence_matrix <- function(x, directions) {
     .Call('_landscapemetrics_rcpp_get_coocurrence_matrix', PACKAGE = 'landscapemetrics', x, directions)
-}
-
-rcpp_get_coocurrence_matrix2 <- function(x, directions) {
-    .Call('_landscapemetrics_rcpp_get_coocurrence_matrix2', PACKAGE = 'landscapemetrics', x, directions)
 }
 
 triangular_index <- function(r, c) {
@@ -96,18 +70,6 @@ rcpp_get_coocurrence_vector <- function(x, directions, ordered = TRUE) {
 
 rcpp_get_offdiagonal_vector <- function(x, directions) {
     .Call('_landscapemetrics_rcpp_get_offdiagonal_vector', PACKAGE = 'landscapemetrics', x, directions)
-}
-
-rcpp_get_entropy <- function(x, base = "log2") {
-    .Call('_landscapemetrics_rcpp_get_entropy', PACKAGE = 'landscapemetrics', x, base)
-}
-
-lsm_unique <- function(x) {
-    .Call('_landscapemetrics_lsm_unique', PACKAGE = 'landscapemetrics', x)
-}
-
-rcpp_get_circle <- function(points, resolution) {
-    .Call('_landscapemetrics_rcpp_get_circle', PACKAGE = 'landscapemetrics', points, resolution)
 }
 
 #' @title First nearest neighbor distance
