--- man/rcpp_get_adjacency.Rd
@@ -1,23 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R
-\name{rcpp_get_adjacency}
-\alias{rcpp_get_adjacency}
-\title{Adjacent cells}
-\usage{
-rcpp_get_adjacency(x, directions)
-}
-\arguments{
-\item{x}{A matrix}
-
-\item{directions}{The number of directions in which cells should be connected:
-4 (rook's case), 8 (queen's case), or a neighbourhood matrix.
-The neighbourhood matrix should have one cell with value 0 (the focal cell),
-and at least one cell with value 1 (the adjacent cells).
-Cells with other values (e.g. NA) are ignored.}
-}
-\description{
-Identify cells that are adjacent to a set of cells on a matrix.
-An output is a two columns matrix, where the first column contains a
-main cell number and the second column contains an adjacent cell number.
-}
-\keyword{internal}
